Move dataframe sorting
